+++ src/Products/MeetingLiege/model/pm_updates.py
@@ -1,7 +1,5 @@
 from collective.datagridcolumns.MultiSelectColumn import MultiSelectColumn
 from Products.Archetypes.atapi import BooleanField
-from Products.Archetypes.atapi import LinesField
-from Products.Archetypes.atapi import MultiSelectionWidget
 from Products.Archetypes.atapi import RichWidget
 from Products.Archetypes.atapi import Schema
 from Products.Archetypes.atapi import SelectionWidget
@@ -13,7 +11,6 @@ from Products.DataGridField import SelectColumn
 from Products.PloneMeeting.config import registerClasses
 from Products.PloneMeeting.config import WriteRiskyConfig
 from Products.PloneMeeting.Meeting import Meeting
-from Products.PloneMeeting.MeetingCategory import MeetingCategory
 from Products.PloneMeeting.MeetingConfig import MeetingConfig
 from Products.PloneMeeting.MeetingItem import MeetingItem
 
@@ -185,32 +182,6 @@ def update_config_schema(baseSchema):
     completeConfigSchema = baseSchema + specificSchema.copy()
     return completeConfigSchema
 MeetingConfig.schema = update_config_schema(MeetingConfig.schema)
-
-
-def update_category_schema(baseSchema):
-
-    specificSchema = Schema((
-        # field for defining a group that is responsible for this category and
-        # that will be able to see every items using this category once this item is validated
-        LinesField(
-            name='groupsOfMatter',
-            widget=MultiSelectionWidget(
-                size=10,
-                description="GroupsOfMatter",
-                description_msgid="groups_of_matter_descr",
-                label='Groupsofmatter',
-                label_msgid='MeetingLiege_label_GroupsOfMatter',
-                i18n_domain='PloneMeeting',
-            ),
-            optional=True,
-            multiValued=1,
-            vocabulary='listGroupsOfMatter',
-        ),
-    ),)
-
-    completeCategorySchema = baseSchema + specificSchema.copy()
-    return completeCategorySchema
-MeetingCategory.schema = update_category_schema(MeetingCategory.schema)
 
 
 registerClasses()